+++ auto_route_generator/lib/src/router_config_visitor.dart
@@ -4,17 +4,24 @@ import 'package:auto_route_generator/route_config_visitor.dart';
 import 'package:auto_route_generator/utils.dart';
 import 'package:source_gen/source_gen.dart';
 
+const _validRouteAnnotations = [
+	'Initial',
+	'MaterialRoute',
+	'CupertinoRoute',
+	'CustomRoute'
+];
+
 // extracts route configs from class fields
-class RouterConfigVisitor extends SimpleElementVisitor {
+class RouterConfigVisitor extends SimpleElementVisitor<RouteConfig> {
   final ConstantReader annotation;
   final routeConfigs = List<RouteConfig>();
 
   RouterConfigVisitor(this.annotation);
 
   @override
-  visitFieldElement(FieldElement field) {
+  RouteConfig visitFieldElement(FieldElement field) {
     final type = field.type;
-    if (type.element is! ClassElement) return;
+    if (type.element is! ClassElement) return null;
 
     final routeConfig = RouteConfig();
     _extractMetaData(field, routeConfig);
@@ -28,9 +35,9 @@ class RouterConfigVisitor extends SimpleElementVisitor {
     routeConfig.className = type.displayName;
 
     final classElement = type.element as ClassElement;
-//     print(    classElement.type.accessors.toString());
-
-    routeConfig.hasWrapper = classElement.allSupertypes.map<String>((el) => el.displayName).contains("AutoRouteWrapper");
+    routeConfig.hasWrapper = classElement.allSupertypes
+		    .map<String>((el) => el.displayName)
+		    .contains('AutoRouteWrapper');
     final constructor = (type.element as ClassElement).unnamedConstructor;
     if (constructor != null && constructor.parameters.isNotEmpty) {
       routeConfig.parameters = constructor.parameters
@@ -39,45 +46,54 @@ class RouterConfigVisitor extends SimpleElementVisitor {
     }
 
     routeConfigs.add(routeConfig);
+    return routeConfig;
   }
 
   void _extractMetaData(FieldElement field, RouteConfig routeConfig) {
     if (field.metadata != null && field.metadata.isNotEmpty) {
-      final autoRoute = field.metadata[0].computeConstantValue();
+	    final autoRoute = field.metadata.first.computeConstantValue();
 
       final type = autoRoute.type.toString();
-      if (type == "Initial") {
+	    // only continue if annotation is MaterialRoute, Initial CupertinoRoute or CustomRoute.
+	    if (!_validRouteAnnotations.contains(type)) {
+		    return;
+	    }
+
+	    // return early if @initial annotation was used
+	    //
+	    // route type will default to MaterialPageRoute
+	    if (type == 'Initial') {
         routeConfig.initial = true;
         return;
-      } else if (type == "MaterialRoute" || type == "CupertinoRoute") {
-        routeConfig.fullscreenDialog =
-            autoRoute.getField("fullscreenDialog")?.toBoolValue();
-        routeConfig.maintainState =
-            autoRoute.getField("maintainState")?.toBoolValue();
-        if (type == "CupertinoRoute") {
-          routeConfig.routeType = RouteType.cupertino;
-          routeConfig.cupertinoNavTitle =
-              autoRoute.getField("title")?.toStringValue();
-        }
-      } else if (type == "CustomRoute") {
+	    }
+
+	    routeConfig.fullscreenDialog =
+			    autoRoute.getField('fullscreenDialog')?.toBoolValue();
+	    routeConfig.maintainState =
+			    autoRoute.getField('maintainState')?.toBoolValue();
+	    routeConfig.initial = autoRoute.getField('initial')?.toBoolValue();
+
+	    if (type == 'CupertinoRoute') {
+		    routeConfig.routeType = RouteType.cupertino;
+		    routeConfig.cupertinoNavTitle =
+				    autoRoute.getField('title')?.toStringValue();
+	    } else if (type == 'CustomRoute') {
         routeConfig.routeType = RouteType.custom;
         routeConfig.durationInMilliseconds =
-            autoRoute.getField("durationInMilliseconds").toIntValue();
-        routeConfig.maintainState =
-            autoRoute.getField("maintainState")?.toBoolValue();
+		        autoRoute.getField('durationInMilliseconds').toIntValue();
         routeConfig.customRouteOpaque =
-            autoRoute.getField("opaque")?.toBoolValue();
+		        autoRoute.getField('opaque')?.toBoolValue();
         routeConfig.customRouteBarrierDismissible =
-            autoRoute.getField("barrierDismissible")?.toBoolValue();
+		        autoRoute.getField('barrierDismissible')?.toBoolValue();
         final function =
-        autoRoute.getField("transitionsBuilder")?.toFunctionValue();
+        autoRoute.getField('transitionsBuilder')?.toFunctionValue();
         if (function != null) {
           final import = getImport(function.source.uri);
           final displayName =
-          function.displayName.replaceFirst(RegExp("^_"), "");
+          function.displayName.replaceFirst(RegExp('^_'), '');
           final functionName = (function.isStatic &&
-              function.enclosingElement?.displayName != null)
-              ? "${function.enclosingElement.displayName}.$displayName"
+		          function.enclosingElement?.displayName != null)
+		          ? '${function.enclosingElement.displayName}.$displayName'
               : displayName;
           routeConfig.transitionBuilder =
               CustomTransitionBuilder(functionName, import);