+++ example/lib/model.dart
@@ -1,0 +1,3 @@
+class Model {
+  String value = "model value";
+}