+++ example/lib/screens/home_screen.dart
@@ -2,16 +2,12 @@ import 'package:auto_route/auto_route_wrapper.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../router.dart';
+import '../model.dart';
 import '../router.gr.dart';
-
 
 class HomeScreen extends StatelessWidget implements AutoRouteWrapper {
   @override
-  Widget get wrappedRoute =>
-      Provider(create: (_) {
-        return Model();
-      }, child: this);
+  Widget get wrappedRoute => Provider(create: (ctx) => Model(), child: this);
 
   @override
   Widget build(BuildContext context) {
@@ -21,9 +17,9 @@ class HomeScreen extends StatelessWidget implements AutoRouteWrapper {
       body: Container(
         child: Center(
           child: FlatButton(
-            child: Text(Provider
-                .of<Model>(context, listen: false)
-                .value),
+	          child: Text(Provider
+			          .of<Model>(context, listen: false)
+			          .value),
             onPressed: () {
               Router.navigator.pushNamed(Router.secondScreenRoute, arguments: SecondScreenArguments(title: "Title"));
             },