+++ example/lib/screens/nested/nested_router.dart
@@ -1,6 +1,7 @@
 import 'package:auto_route/auto_route_annotations.dart';
 import 'package:example/screens/nested/nested_home_page.dart';
-import 'package:example/screens/nested/nested_second_page.dart';
+
+import 'nested_second_page.dart';
 
 @autoRouter
 class $NestedRouter {